code refactor and spec fixes
