+++ spec/operations/transmittable/add_error_spec.rb
@@ -3,7 +3,7 @@
 require 'rails_helper'
 require 'shared_examples/ssa_transmittable'
 
-RSpec.describe Jobs::AddError, dbclean: :after_each do
+RSpec.describe Transmittable::AddError, dbclean: :after_each do
   include_context "ssa transmittable job transmission transaction"
   subject { described_class.new }
